Centralize custom icon registration and disable in CSE
CMK-14279
